Rename WFST to DrawFeature
