+++ docs/SUMMARY.md
@@ -13,6 +13,7 @@ Boundless SDK
     * [Bookmarks](api/Bookmarks.md)
     * [Button](api/Button.md)
     * [Chart](api/Chart.md)
+    * [DrawFeature](api/DrawFeature.md)
     * [EditPopup](api/EditPopup.md)
     * [FeatureTable](api/FeatureTable.md)
     * [FillEditor](api/FillEditor.md)
@@ -52,7 +53,6 @@ Boundless SDK
     * [Select](api/Select.md)
     * [StrokeEditor](api/StrokeEditor.md)
     * [StyleModal](api/StyleModal.md)
-    * [WFST](api/WFST.md)
     * [WMSLegend](api/WMSLegend.md)
     * [Zoom](api/Zoom.md)
     * [ZoomSlider](api/ZoomSlider.md)
